Rename NarrativeElements package for consistency
This commit renames the routes/library/NarrativeElements package to
src/routes/library/narrative for consistency with other package names.
This also includes the modules contained in this package.
